+++ go/cmd/server.go
@@ -23,6 +23,7 @@ import (
 var ServerCmd = &cobra.Command{
 	Use: "server",
 	Run: func(cmd *cobra.Command, args []string) {
+		server.ServerFlags.DataDirectory = dataDirectory
 		server.ServerMain()
 	},
 }
@@ -31,7 +32,6 @@ func init() {
 	flags := ServerCmd.Flags()
 	flags.Int16VarP(&server.ServerFlags.Port, "port", "p", 6045, "Port")
 	flags.StringVar(&server.ServerFlags.Host, "host", "127.0.0.1", "Host to bind to")
-	flags.StringVar(&server.ServerFlags.LogFilename, "log", "maker.log", "Log filename")
 	flags.BoolVar(&server.ServerFlags.NoLog, "nolog", false, "Disable logging to file")
 	flags.BoolVar(&server.ServerFlags.OpenBrowser, "open", false, "Open browser")
 	rootCmd.AddCommand(ServerCmd)